+++ src/file_pipe_log/pipe.rs
@@ -2,7 +2,7 @@
 
 use std::collections::VecDeque;
 use std::fs::File as StdFile;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::Arc;
 
 use crossbeam::utils::CachePadded;
@@ -12,6 +12,7 @@ use parking_lot::{Mutex, MutexGuard, RwLock};
 
 use crate::config::Config;
 use crate::env::FileSystem;
+use crate::errors::is_no_space_err;
 use crate::event_listener::EventListener;
 use crate::metrics::*;
 use crate::pipe_log::{
@@ -23,13 +24,14 @@ use super::format::{build_recycled_file_name, FileNameExt, LogFileFormat};
 use super::log_file::build_file_reader;
 use super::log_file::{build_file_writer, LogFileWriter};
 
+pub type PathId = usize;
+pub type Paths = Vec<PathBuf>;
+
+/// Main directory path id.
 pub const DEFAULT_PATH_ID: PathId = 0;
 /// FileSeq of logs must start from `1` by default to keep backward
 /// compatibility.
 pub const DEFAULT_FIRST_FILE_SEQ: FileSeq = 1;
-
-pub type PathId = usize;
-pub type Paths = Vec<PathBuf>;
 
 #[derive(Debug)]
 pub struct File<F: FileSystem> {
@@ -78,13 +80,13 @@ impl<F: FileSystem> SinglePipe<F> {
     /// Opens a new [`SinglePipe`].
     pub fn open(
         cfg: &Config,
+        paths: Paths,
         file_system: Arc<F>,
         listeners: Vec<Arc<dyn EventListener>>,
         queue: LogQueue,
         mut active_files: Vec<File<F>>,
         recycled_files: Vec<File<F>>,
     ) -> Result<Self> {
-        let paths = vec![Path::new(&cfg.dir).to_path_buf()];
         let alignment = || {
             fail_point!("file_pipe_log::open::force_set_alignment", |_| { 16 });
             0
@@ -94,7 +96,7 @@ impl<F: FileSystem> SinglePipe<F> {
         // Open or create active file.
         let no_active_files = active_files.is_empty();
         if no_active_files {
-            let path_id = DEFAULT_PATH_ID;
+            let path_id = find_available_dir(&paths, cfg.target_file_size.0 as usize);
             let file_id = FileId::new(queue, DEFAULT_FIRST_FILE_SEQ);
             let path = file_id.build_file_path(&paths[path_id]);
             active_files.push(File {
@@ -147,10 +149,9 @@ impl<F: FileSystem> SinglePipe<F> {
 
     /// Synchronizes all metadatas associated with the working directory to the
     /// filesystem.
-    fn sync_dir(&self) -> Result<()> {
+    fn sync_dir(&self, path_id: PathId) -> Result<()> {
         debug_assert!(!self.paths.is_empty());
-        let path = PathBuf::from(&self.paths[0]);
-        std::fs::File::open(path).and_then(|d| d.sync_all())?;
+        std::fs::File::open(&PathBuf::from(&self.paths[path_id])).and_then(|d| d.sync_all())?;
         Ok(())
     }
 
@@ -171,7 +172,7 @@ impl<F: FileSystem> SinglePipe<F> {
                 return Ok((f.path_id, self.file_system.open(&dst_path)?));
             }
         }
-        let path_id = DEFAULT_PATH_ID;
+        let path_id = find_available_dir(&self.paths, self.target_file_size);
         let dst_path = new_file_id.build_file_path(&self.paths[path_id]);
         Ok((path_id, self.file_system.create(&dst_path)?))
     }
@@ -218,7 +219,7 @@ impl<F: FileSystem> SinglePipe<F> {
         // File header must be persisted. This way we can recover gracefully if power
         // loss before a new entry is written.
         new_file.writer.sync()?;
-        self.sync_dir()?;
+        self.sync_dir(path_id)?;
 
         **writable_file = new_file;
         let len = {
@@ -302,7 +303,30 @@ impl<F: FileSystem> SinglePipe<F> {
                     seq, e, te
                 );
             }
-            return Err(e);
+            if is_no_space_err(&e) {
+                // TODO: There exists several corner cases should be tackled if
+                // `bytes.len()` > `target_file_size`. For example,
+                // - [1] main-dir has no recycled logs, and spill-dir have several recycled
+                //   logs.
+                // - [2] main-dir has several recycled logs, and sum(recycled_logs.size()) <
+                //   expected_file_size, but no recycled logs exist in spill-dir.
+                // - [3] Both main-dir and spill-dir have several recycled logs.
+                // But as `bytes.len()` is always smaller than `target_file_size` in common
+                // cases, this issue will be ignored temprorarily.
+                if let Err(e) = self.rotate_imp(&mut writable_file) {
+                    panic!(
+                        "error when rotate [{:?}:{}]: {}",
+                        self.queue, writable_file.seq, e
+                    );
+                }
+                // If there still exists free space for this record, rotate the file
+                // and return a special TryAgain Err (for retry) to the caller.
+                return Err(Error::TryAgain(format!(
+                    "error when append [{:?}:{}]: {}",
+                    self.queue, seq, e
+                )));
+            }
+            return Err(Error::Io(e));
         }
         let handle = FileBlockHandle {
             id: FileId {
@@ -411,14 +435,14 @@ impl<F: FileSystem> SinglePipe<F> {
 pub struct DualPipes<F: FileSystem> {
     pipes: [SinglePipe<F>; 2],
 
-    _dir_lock: StdFile,
+    _dir_locks: Vec<StdFile>,
 }
 
 impl<F: FileSystem> DualPipes<F> {
     /// Open a new [`DualPipes`]. Assumes the two [`SinglePipe`]s share the
     /// same directory, and that directory is locked by `dir_lock`.
     pub(super) fn open(
-        dir_lock: StdFile,
+        dir_locks: Vec<StdFile>,
         appender: SinglePipe<F>,
         rewriter: SinglePipe<F>,
     ) -> Result<Self> {
@@ -428,7 +452,7 @@ impl<F: FileSystem> DualPipes<F> {
 
         Ok(Self {
             pipes: [appender, rewriter],
-            _dir_lock: dir_lock,
+            _dir_locks: dir_locks,
         })
     }
 
@@ -479,8 +503,27 @@ impl<F: FileSystem> PipeLog for DualPipes<F> {
     }
 }
 
+/// Fetch and return a valid `PathId` of the specific directories.
+pub(crate) fn find_available_dir(paths: &Paths, target_size: usize) -> PathId {
+    fail_point!("file_pipe_log::force_choose_dir", |s| s
+        .map_or(DEFAULT_PATH_ID, |n| n.parse::<usize>().unwrap()));
+    // Only if one single dir is set by `Config::dir`, can it skip the check of disk
+    // space usage.
+    if paths.len() > 1 {
+        for (t, p) in paths.iter().enumerate() {
+            if let Ok(disk_stats) = fs2::statvfs(&p) {
+                if target_size <= disk_stats.available_space() as usize {
+                    return t;
+                }
+            }
+        }
+    }
+    DEFAULT_PATH_ID
+}
+
 #[cfg(test)]
 mod tests {
+    use std::path::Path;
     use tempfile::Builder;
 
     use super::super::format::LogFileFormat;
@@ -492,17 +535,28 @@ mod tests {
 
     fn new_test_pipe<F: FileSystem>(
         cfg: &Config,
+        paths: Paths,
         queue: LogQueue,
         fs: Arc<F>,
     ) -> Result<SinglePipe<F>> {
-        SinglePipe::open(cfg, fs, Vec::new(), queue, Vec::new(), Vec::new())
+        SinglePipe::open(cfg, paths, fs, Vec::new(), queue, Vec::new(), Vec::new())
     }
 
     fn new_test_pipes(cfg: &Config) -> Result<DualPipes<DefaultFileSystem>> {
         DualPipes::open(
-            lock_dir(&cfg.dir)?,
-            new_test_pipe(cfg, LogQueue::Append, Arc::new(DefaultFileSystem))?,
-            new_test_pipe(cfg, LogQueue::Rewrite, Arc::new(DefaultFileSystem))?,
+            vec![lock_dir(&cfg.dir)?],
+            new_test_pipe(
+                cfg,
+                vec![Path::new(&cfg.dir).to_path_buf()],
+                LogQueue::Append,
+                Arc::new(DefaultFileSystem),
+            )?,
+            new_test_pipe(
+                cfg,
+                vec![Path::new(&cfg.dir).to_path_buf()],
+                LogQueue::Rewrite,
+                Arc::new(DefaultFileSystem),
+            )?,
         )
     }
 
@@ -613,7 +667,8 @@ mod tests {
         };
         let queue = LogQueue::Append;
         let fs = Arc::new(ObfuscatedFileSystem::default());
-        let pipe_log = new_test_pipe(&cfg, queue, fs).unwrap();
+        let pipe_log =
+            new_test_pipe(&cfg, vec![Path::new(&cfg.dir).to_path_buf()], queue, fs).unwrap();
         assert_eq!(pipe_log.file_span(), (1, 1));
 
         fn content(i: usize) -> Vec<u8> {